Remove compile errors on T4.x
There are a couple wasys of doing it.  Your code was comparing an address to a uint32_t value... So I simply did cast of the address to uint32_t...

Other option would be to change your definitions of things like FLASH_START to simply be a uint32_t.

Warning it also include a change to .gitignore as to keep changes brought in by VisualMicro to being uploaded to project